add buildSrc
