--- .idea/misc.xml
@@ -1,67 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project version="4">
-  <component name="DesignSurface">
-    <option name="filePathToZoomLevelMap">
-      <map>
-        <entry key="../../../../layout/compose-model-1615095442020.xml" value="1.0" />
-        <entry key="../../../../layout/compose-model-1615126828677.xml" value="1.1514815450127516" />
-        <entry key="../../../../layout/compose-model-1615135671005.xml" value="0.5" />
-        <entry key="../../../../layout/compose-model-1615137650472.xml" value="0.75" />
-        <entry key="../../../../layout/compose-model-1615306012206.xml" value="0.41069887834339946" />
-        <entry key="../../../../layout/compose-model-1615306308911.xml" value="0.41069887834339946" />
-        <entry key="../../../../layout/compose-model-1615306798824.xml" value="0.22" />
-        <entry key="../../../../layout/compose-model-1615307186826.xml" value="0.4253666954270923" />
-        <entry key="../../../../layout/compose-model-1615308035636.xml" value="0.225" />
-        <entry key="../../../../layout/compose-model-1615308948575.xml" value="0.826629680998613" />
-        <entry key="../../../../layout/compose-model-1615308998635.xml" value="0.4253666954270923" />
-        <entry key="../../../../layout/compose-model-1615309012551.xml" value="0.20819256756756757" />
-        <entry key="../../../../layout/compose-model-1615309338338.xml" value="0.4253666954270923" />
-        <entry key="../../../../layout/compose-model-1615310005413.xml" value="0.33" />
-        <entry key="../../../../layout/compose-model-1615310597661.xml" value="0.4201898188093184" />
-        <entry key="../../../../layout/compose-model-1615311638518.xml" value="0.5" />
-        <entry key="../../../../layout/compose-model-1615311734279.xml" value="0.5176876617773943" />
-        <entry key="../../../../layout/compose-model-1615311738571.xml" value="0.7919556171983356" />
-        <entry key="../../../../layout/compose-model-1615313326602.xml" value="0.7919556171983356" />
-        <entry key="../../../../layout/compose-model-1615313340653.xml" value="0.7919556171983356" />
-        <entry key="../../../../layout/compose-model-1615313441896.xml" value="0.7919556171983356" />
-        <entry key="../../../../layout/compose-model-1615313460935.xml" value="0.7919556171983356" />
-        <entry key="../../../../layout/compose-model-1615313734724.xml" value="1.0" />
-        <entry key="../../../../layout/compose-model-1615315532170.xml" value="0.7919556171983356" />
-        <entry key="../../../../layout/compose-model-1615315716841.xml" value="0.19578059071729959" />
-        <entry key="../../../../layout/compose-model-1615315792950.xml" value="0.38861151727973925" />
-        <entry key="../../../../layout/compose-model-1615316567324.xml" value="0.7193798449612403" />
-        <entry key="../../../../layout/compose-model-1615469898962.xml" value="0.2677364864864865" />
-        <entry key="../../../../layout/compose-model-1615469960966.xml" value="0.772093023255814" />
-        <entry key="../../../../layout/compose-model-1615469981414.xml" value="0.5" />
-        <entry key="../../../../layout/compose-model-1615565344846.xml" value="0.4046296296296296" />
-        <entry key="../../../../layout/compose-model-1615643782229.xml" value="0.4305555555555556" />
-        <entry key="../../../../layout/compose-model-1615644068833.xml" value="0.6449375866851595" />
-        <entry key="../../../../layout/compose-model-1615644300162.xml" value="0.38263229308005425" />
-        <entry key="../../../../layout/compose-model-1615644936104.xml" value="0.6324549237170597" />
-        <entry key="../../../../layout/compose-model-1615646867366.xml" value="0.6324549237170597" />
-        <entry key="../../../../layout/compose-model-1615647044684.xml" value="0.38159675236806495" />
-        <entry key="../../../../layout/compose-model-1615647058407.xml" value="0.1638888888888889" />
-        <entry key="../../../../layout/compose-model-1615647573279.xml" value="0.2829403606102635" />
-        <entry key="../../../../layout/compose-model-1615661794701.xml" value="0.296875" />
-        <entry key="../../../../layout/compose-model-1615662260308.xml" value="0.75" />
-        <entry key="../../../../layout/compose-model-1615662560400.xml" value="0.5454545454545454" />
-        <entry key="../../../../layout/compose-model-1615663523132.xml" value="0.5454545454545454" />
-        <entry key="../../../../layout/compose-model-1615663551495.xml" value="0.5454545454545454" />
-        <entry key="../../../../layout/compose-model-1615663673450.xml" value="0.5454545454545454" />
-        <entry key="../../../../layout/compose-model-1615664472539.xml" value="0.46421663442940037" />
-        <entry key="../../../../layout/compose-model-1615664494149.xml" value="0.5490375008656208" />
-        <entry key="../../../../layout/compose-model-1615664506924.xml" value="0.5454545454545454" />
-        <entry key="../../../../layout/compose-model-1615665158259.xml" value="0.5454545454545454" />
-        <entry key="../../../../layout/compose-model-1615729666041.xml" value="0.23417721518987342" />
-        <entry key="app/src/main/res/layout/activity_main.xml" value="0.23020833333333332" />
-      </map>
-    </option>
-  </component>
-  <component name="ExternalStorageConfigurationManager" enabled="true" />
-  <component name="ProjectRootManager" version="2" languageLevel="JDK_11" default="true" project-jdk-name="JDK" project-jdk-type="JavaSDK">
-    <output url="file://$PROJECT_DIR$/build/classes" />
-  </component>
-  <component name="ProjectType">
-    <option name="id" value="Android" />
-  </component>
-</project>
