+++ app/build.gradle
@@ -1,3 +1,6 @@
+import com.example.androidcomposesample.buildsrc.Versions
+import com.example.androidcomposesample.buildsrc.Libs
+
 plugins {
     id 'com.android.application'
     id 'kotlin-android'
@@ -9,7 +12,7 @@ android {
 
     defaultConfig {
         applicationId "com.example.androidcomposesample"
-        minSdkVersion 21
+        minSdkVersion 23
         targetSdkVersion 30
         versionCode 1
         versionName "1.0"
@@ -34,37 +37,35 @@ android {
         compose true
     }
     composeOptions {
-        kotlinCompilerVersion '1.4.21'
-        kotlinCompilerExtensionVersion compose_version
+        kotlinCompilerVersion Versions.kotlin
+        kotlinCompilerExtensionVersion Versions.compose
     }
 }
 
 dependencies {
-
-    implementation 'androidx.core:core-ktx:1.3.2'
-    implementation 'androidx.appcompat:appcompat:1.2.0'
-    implementation 'com.google.android.material:material:1.3.0'
-    implementation 'androidx.constraintlayout:constraintlayout:2.0.4'
-    testImplementation 'junit:junit:4.13.2'
-    androidTestImplementation 'androidx.test.ext:junit:1.1.2'
-    androidTestImplementation 'androidx.test.espresso:espresso-core:3.3.0'
+    implementation(Libs.Kotlin.stdlib)
+    implementation(Libs.AndroidX.coreKtx)
+    implementation(Libs.AndroidX.appcompat)
+    implementation(Libs.AndroidX.ConstraintLayout.core)
 
     //navigation
-    implementation "androidx.navigation:navigation-compose:1.0.0-alpha09"
+    implementation(Libs.AndroidX.navigation)
 
-    implementation "androidx.compose.ui:ui:$compose_version"
-    implementation "androidx.compose.ui:ui-tooling:$compose_version"
-    implementation "androidx.compose.foundation:foundation-layout:$compose_version"
-    implementation "androidx.compose.material:material:$compose_version"
-    implementation "androidx.compose.material:material-icons-core:$compose_version"
-    implementation "androidx.compose.material:material-icons-extended:$compose_version"
-    implementation "androidx.compose.runtime:runtime-livedata:$compose_version"
-
+    implementation(Libs.AndroidX.Compose.ui)
+    implementation(Libs.AndroidX.Compose.tooling)
+    implementation(Libs.AndroidX.Compose.layout)
+    implementation(Libs.AndroidX.Compose.material)
+    implementation(Libs.AndroidX.Compose.materialIcon)
+    implementation(Libs.AndroidX.Compose.ui)
+    implementation(Libs.AndroidX.Compose.ui)
+    implementation(Libs.AndroidX.Compose.ui)
+    implementation(Libs.AndroidX.Compose.runtime)
+    implementation(Libs.AndroidX.Activity.activityCompose)
     //constraintLayout
-    implementation "androidx.constraintlayout:constraintlayout-compose:1.0.0-alpha03"
+    implementation(Libs.AndroidX.ConstraintLayout.compose)
+    implementation(Libs.material)
 
     //coil
     implementation "dev.chrisbanes.accompanist:accompanist-coil:0.6.1"
 
-    implementation "androidx.activity:activity-compose:1.3.0-alpha04"
 }
