+++ app/src/main/java/com/example/androidcomposesample/todo/DataGenerators.kt
@@ -1,4 +1,4 @@
-package com.example.androidcomposesample.todp
+package com.example.androidcomposesample.todo
 
 fun generateRandomTodoItem(): TodoItem {
     val message = listOf(
