+++ build.gradle
@@ -1,16 +1,21 @@
+import com.example.androidcomposesample.buildsrc.Versions
+import com.example.androidcomposesample.buildsrc.Libs
+
 buildscript {
-    ext.kotlin_version = "1.4.30"
-    ext.compose_version = '1.0.0-beta02'
     repositories {
         google()
+        jcenter()
         mavenCentral()
     }
     dependencies {
         classpath 'com.android.tools.build:gradle:7.0.0-alpha14'
-        classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:1.4.32"
+        classpath Libs.Kotlin.stdlib
+        classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:${Versions.kotlin}"
     }
 }
-
-task clean(type: Delete) {
-    delete rootProject.buildDir
+allprojects {
+    repositories {
+        google()
+        jcenter()
+    }
 }
