--- buildSrc/build.gradle.kt
@@ -1,7 +1,0 @@
-repositories {
-    jcenter()
-}
-
-plugins {
-    `kotlin-dsl`
-}
