+++ buildSrc/src/main/java/Dependencies.kt
@@ -1,0 +1,79 @@
+package com.example.androidcomposesample.buildsrc
+
+object Versions {
+    const val kotlin = "1.4.32"
+    const val coroutines = "1.4.1"
+    const val compose = "1.0.0-beta04"
+    const val ktlint = "0.40.0"
+}
+
+object Libs {
+    const val androidGradlePlugin = "com.android.tools.build:gradle:7.0.0-alpha08"
+
+    object Accompanist {
+        private const val version = "0.6.0"
+        const val coil = "dev.chrisbanes.accompanist:accompanist-coil:$version"
+        const val insets = "dev.chrisbanes.accompanist:accompanist-insets:$version"
+    }
+
+    object Kotlin {
+        const val stdlib = "org.jetbrains.kotlin:kotlin-stdlib-jdk7:${Versions.kotlin}"
+        const val gradlePlugin = "org.jetbrains.kotlin:kotlin-gradle-plugin:${Versions.kotlin}"
+        const val extensions = "org.jetbrains.kotlin:kotlin-android-extensions:${Versions.kotlin}"
+    }
+
+    object Coroutines {
+        const val core = "org.jetbrains.kotlinx:kotlinx-coroutines-core:${Versions.coroutines}"
+        const val android = "org.jetbrains.kotlinx:kotlinx-coroutines-android:${Versions.coroutines}"
+        const val test = "org.jetbrains.kotlinx:kotlinx-coroutines-test:${Versions.coroutines}"
+    }
+
+    object JUnit {
+        private const val version = "4.13"
+        const val junit = "junit:junit:$version"
+    }
+
+    object AndroidX {
+        const val coreKtx = "androidx.core:core-ktx:1.5.0-beta01"
+        const val appcompat = "androidx.appcompat:appcompat:1.2.0"
+        const val navigation = "androidx.navigation:navigation-compose:1.0.0-alpha10"
+
+        object Compose {
+            const val animation = "androidx.compose.animation:animation:${Versions.compose}"
+            const val foundation = "androidx.compose.foundation:foundation:${Versions.compose}"
+            const val layout = "androidx.compose.foundation:foundation-layout:${Versions.compose}"
+            const val material = "androidx.compose.material:material:${Versions.compose}"
+            const val materialIcon = "androidx.compose.material:material-icons-extended:${Versions.compose}"
+            const val runtime = "androidx.compose.runtime:runtime:${Versions.compose}"
+            const val tooling = "androidx.compose.ui:ui-tooling:${Versions.compose}"
+            const val ui = "androidx.compose.ui:ui:${Versions.compose}"
+            const val uiUtil = "androidx.compose.ui:ui-util:${Versions.compose}"
+            const val uiTest = "androidx.compose.ui:ui-test-junit4:${Versions.compose}"
+        }
+
+        object Activity {
+            const val activityCompose = "androidx.activity:activity-compose:1.3.0-alpha06"
+        }
+
+        object ConstraintLayout {
+            const val core = "androidx.constraintlayout:constraintlayout:2.0.4"
+            const val compose = "androidx.constraintlayout:constraintlayout-compose:1.0.0-alpha05"
+        }
+
+        object Test {
+            private const val version = "1.3.0"
+            const val core = "androidx.test:core:$version"
+            const val rules = "androidx.test:rules:$version"
+
+            object Ext {
+                private const val version = "1.1.2"
+                const val junit = "androidx.test.ext:junit-ktx:$version"
+            }
+
+            const val espressoCore = "androidx.test.espresso:espresso-core:3.3.0"
+        }
+    }
+
+    const val material = "com.google.android.material:material:1.3.0"
+}
+
