--- buildSrc/src/main/java/com/example/androidcomposesample/buildsrc/Dependencies.kt
@@ -1,78 +1,0 @@
-package com.example.androidcomposesample.buildsrc
-
-object Versions {
-    const val ktlint = "0.40.0"
-}
-
-object Libs {
-    const val androidGradlePlugin = "com.android.tools.build:gradle:7.0.0-alpha08"
-
-    object Accompanist {
-        private const val version = "0.6.0"
-        const val coil = "dev.chrisbanes.accompanist:accompanist-coil:$version"
-        const val insets = "dev.chrisbanes.accompanist:accompanist-insets:$version"
-    }
-
-    object Kotlin {
-        private const val version = "1.4.30"
-        const val stdlib = "org.jetbrains.kotlin:kotlin-stdlib-jdk8:$version"
-        const val gradlePlugin = "org.jetbrains.kotlin:kotlin-gradle-plugin:$version"
-        const val extensions = "org.jetbrains.kotlin:kotlin-android-extensions:$version"
-    }
-
-    object Coroutines {
-        private const val version = "1.4.1"
-        const val core = "org.jetbrains.kotlinx:kotlinx-coroutines-core:$version"
-        const val android = "org.jetbrains.kotlinx:kotlinx-coroutines-android:$version"
-        const val test = "org.jetbrains.kotlinx:kotlinx-coroutines-test:$version"
-    }
-
-    object JUnit {
-        private const val version = "4.13"
-        const val junit = "junit:junit:$version"
-    }
-
-    object AndroidX {
-        const val coreKtx = "androidx.core:core-ktx:1.5.0-beta01"
-        const val navigation = "androidx.navigation:navigation-compose:1.0.0-alpha08"
-
-        object Compose {
-            const val snapshot = ""
-            const val version = "1.0.0-beta01"
-
-            const val animation = "androidx.compose.animation:animation:$version"
-            const val foundation = "androidx.compose.foundation:foundation:$version"
-            const val layout = "androidx.compose.foundation:foundation-layout:$version"
-            const val iconsExtended = "androidx.compose.material:material-icons-extended:$version"
-            const val material = "androidx.compose.material:material:$version"
-            const val runtime = "androidx.compose.runtime:runtime:$version"
-            const val tooling = "androidx.compose.ui:ui-tooling:$version"
-            const val ui = "androidx.compose.ui:ui:$version"
-            const val uiUtil = "androidx.compose.ui:ui-util:$version"
-            const val uiTest = "androidx.compose.ui:ui-test-junit4:$version"
-        }
-
-        object Activity {
-            const val activityCompose = "androidx.activity:activity-compose:1.3.0-alpha03"
-        }
-
-        object ConstraintLayout {
-            const val constraintLayoutCompose =
-                "androidx.constraintlayout:constraintlayout-compose:1.0.0-alpha03"
-        }
-
-        object Test {
-            private const val version = "1.3.0"
-            const val core = "androidx.test:core:$version"
-            const val rules = "androidx.test:rules:$version"
-
-            object Ext {
-                private const val version = "1.1.2"
-                const val junit = "androidx.test.ext:junit-ktx:$version"
-            }
-
-            const val espressoCore = "androidx.test.espresso:espresso-core:3.3.0"
-        }
-    }
-}
-
